Added a bunch of initial game engine functions
added a function to check if the current game status is a win,
tie, or still in progress. Added a bunch of functions that I will
potentially keep but not sure yet including functions to push
'X' or 'O' to the game board array, an event listener.  Created
a new events.js file.  Didn't know we were supposed to use branches
until I was already deep in the project but will use different
'feature' branches in the future.