Added more game logic
added a function that checks for a winner, tie, and whether or not
a next turn should be expected.
Also added a function that pushes the current player's move to an
array, then switches the player and console logs who's turn it is.
This function also prevents a player from duplicating a space.